make do_gettimeofday() inline
do_gettimeofday() are only used by a few modules now any
more (waiting for the respective patches to get accepted), and they are
among the last holdouts of code that is not y2038 safe in the core kernel.

Move the implementation into the timekeeping.h header to clean up
the core kernel and isolate the old interfaces further.

also to support KLAPSE modules.

Signed-off-by: Rudi Setiyawan <diphons@gmail.com>